spond-whatsapp/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
spond==1.1.1
sqlite3